backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "google-auth>=2.40.3",
    "groq>=0.31.0",
    "lxml>=6.0.0",
    "psycopg>=3.2.9",
    "psycopg-pool>=3.2.6",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.8",
]
